requirements.txt
mysql-connector-python~=8.3.0
bcrypt
mysql-connector-python
